+++ CHANGELOG.md
@@ -5,6 +5,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* Filtering (whitelist and blacklist) of what is indexed in Search, based on Google's Common Expression Language.  See [details here](./search/README.md). Added `--search-common-action-filter-on-expr` and `--search-common-action-filter-out-expr`.
+    * NOTE: This doesn't affect what is extracted from the chain, allowing you to re-index selectively without a chain replay.
+
+
+### Changed
+* CLI: `--search-indexer-dfuse-hooks-action-name`, `--search-live-dfuse-hooks-action-name`, `--search-forkresolver-dfuse-hooks-action-name` fused into new `--search-common-dfuse-hooks-action-name`.
+* CLI: Similarly, `--search-...-mesh-publish-polling-duration` fused into new `--search-common-mesh-publish-polling-duration`.
+* CLI: Similarly, all of the `--search-mesh-...` options were renamed to `--search-common-mesh-...` (previously `--search-mesh-service-version`, `--search-mesh-namespace`, `--search-mesh-store-addr`)
+
+### Removed
+* The `--search-...-indexing-restrictions-json`.  This was replaced by the filtering listed above.
+
 
 
 ## [v0.1.0-beta1] 2020-04-17